+++ src/components/EventAdmin/EventAdminQueue/EventAdminQueue.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import mapStoreToProps from '../../redux/mapStoreToProps';
+import mapStoreToProps from '../../../redux/mapStoreToProps';
 
 // Basic class component structure for React with default state
 // value setup. When making a new component be sure to replace
 // the component name TemplateClass with the name for the new
 // component.
-class UserMaterialRequest extends Component {
+class EventAdminQueue extends Component {
     state = {
         heading: 'Class Component',
     };
@@ -20,4 +20,4 @@ class UserMaterialRequest extends Component {
     }
 }
 
-export default connect(mapStoreToProps)(UserMaterialRequest);
+export default connect(mapStoreToProps)(EventAdminQueue);
