+++ src/components/User/UserDrawingSubmit/UserDrawingSubmit.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import mapStoreToProps from '../../redux/mapStoreToProps';
+import mapStoreToProps from '../../../redux/mapStoreToProps';
 
 // Basic class component structure for React with default state
 // value setup. When making a new component be sure to replace
 // the component name TemplateClass with the name for the new
 // component.
-class EventAdminPending extends Component {
+class UserDrawingSubmit extends Component {
     state = {
         heading: 'Class Component',
     };
@@ -20,4 +20,4 @@ class EventAdminPending extends Component {
     }
 }
 
-export default connect(mapStoreToProps)(EventAdminPending);
+export default connect(mapStoreToProps)(UserDrawingSubmit);
